requirements/base.txt
psycopg2-binary==2.9.5

kafka-python==2.0.2

pydantic==1.10.4
typing_extensions==4.4.0

click==8.1.3
typer==0.7.0
certifi==2022.12.7
colorama==0.4.6
commonmark==0.9.1
Pygments==2.14.0
rich==12.6.0
shellingham==1.5.0.post1

certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
urllib3==1.26.14
requests==2.28.2

python-json-logger==2.0.4
